+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- shadow-rs long_version.
+- Windows version resource & application manifest.
+
 ## 0.1.2 - 2022-10-09
 ### Added
 - Prebuilt binaries.